requirements.txt
Django>=5.0,<6.0
aiogram>=3.13.0
python-dotenv>=1.0.0
